refactor(slack): Extract MissingScopeError to separate errors module
- Move MissingScopeError class from messaging.ts to dedicated errors.ts file
- Improve module organization by separating error types from utilities
- Update messaging.ts to import and re-export MissingScopeError for backwards compatibility
- Export error module from slack module index

Rationale: As the Slack integration grows, organizing error types in a dedicated module improves maintainability and makes it easier to discover and reuse error types across the codebase. This separation follows the single-responsibility principle and makes the messaging module more focused on its core functionality.
