+++ camille-bot/src/shared/slack/index.ts
@@ -1,9 +1,10 @@
 /**
  * Slack Module
- * 
+ *
  * Exports all Slack-related functions and types
  */
 
+export * from './errors';
 export * from './events';
 export * from './utils';
 export * from './types';
